Final Iteration

+++ Jio/Models/Restaurant.cs
@@ -12,7 +12,7 @@ namespace Jio.Models
         public string Name { get; set; }
         public string Description { get; set; }
         public string Owner { get; set; }
-        public int Contact { get; set; }
+        public double Contact { get; set; }
         public String address { get; set; }
         public virtual ICollection<Menu_Item> menu_item { get; set; }
     }
